Flattened linode_api.objects.nodebalancer into a file
