+++ linode_api/objects/account.py
@@ -5,7 +5,7 @@ import requests
 from linode_api.errors import ApiError, UnexpectedResponseError
 from linode_api.objects import (Base, DerivedBase, Domain, Image, Linode,
                                 Property, StackScript, Volume)
-from linode_api.objects.nodebalancer.nodebalancer import NodeBalancer
+from linode_api.objects.nodebalancer import NodeBalancer
 from linode_api.objects.support.ticket import SupportTicket
 from linode_api.objects.longview import LongviewClient, LongviewSubscription
