+++ linode_api/objects/nodebalancer.py
@@ -3,9 +3,43 @@ from __future__ import absolute_import
 import os
 
 from linode_api.errors import UnexpectedResponseError
-from linode_api.objects import DerivedBase, Property
 
-from .node import NodeBalancerNode
+from linode_api.objects import Base, DerivedBase, Property, Region
+from linode_api.objects.networking import IPAddress
+
+
+class NodeBalancerNode(DerivedBase):
+    api_endpoint = '/nodebalancers/{nodebalancer_id}/configs/{config_id}/nodes/{id}'
+    derived_url_path = 'nodes'
+    parent_id_name='config_id'
+
+    properties = {
+        'id': Property(identifier=True),
+        'config_id': Property(identifier=True),
+        'nodebalancer_id': Property(identifier=True),
+        "label": Property(mutable=True),
+        "address": Property(mutable=True),
+        "weight": Property(mutable=True),
+        "mode": Property(mutable=True),
+        "status": Property(),
+    }
+
+    def __init__(self, client, id, parent_id, nodebalancer_id=None, json=None):
+        """
+        We need a special constructor here because this object's parent
+        has a parent itself.
+        """
+        if not nodebalancer_id and not isinstance(parent_id, tuple):
+            raise ValueError('NodeBalancerNode must either be created with a nodebalancer_id or a tuple of '
+                    '(config_id, nodebalancer_id) for parent_id!')
+
+        if isinstance(parent_id, tuple):
+            nodebalancer_id = parent_id[1]
+            parent_id = parent_id[0]
+
+        DerivedBase.__init__(self, client, id, parent_id, json=json)
+
+        self._set('nodebalancer_id', nodebalancer_id)
 
 
 class NodeBalancerConfig(DerivedBase):
@@ -90,3 +124,35 @@ class NodeBalancerConfig(DerivedBase):
         if os.path.isfile(os.path.expanduser(key_file)):
             with open(os.path.expanduser(key_file)) as f:
                 self.ssl_key = f.read() # pylint: disable=attribute-defined-outside-init
+
+
+class NodeBalancer(Base):
+    api_endpoint = '/nodebalancers/{id}'
+    properties = {
+        'id': Property(identifier=True),
+        'label': Property(mutable=True),
+        'hostname': Property(),
+        'client_conn_throttle': Property(mutable=True),
+        'status': Property(),
+        'created': Property(is_datetime=True),
+        'updated': Property(is_datetime=True),
+        'ipv4': Property(relationship=IPAddress),
+        'ipv6': Property(),
+        'region': Property(slug_relationship=Region, filterable=True),
+        'configs': Property(derived_class=NodeBalancerConfig),
+    }
+
+    # create derived objects
+    def config_create(self, label=None, **kwargs):
+        params = kwargs
+        if label:
+            params['label'] = label
+
+        result = self._client.post("{}/configs".format(NodeBalancer.api_endpoint), model=self, data=params)
+        self.invalidate()
+
+        if not 'id' in result:
+            raise UnexpectedResponseError('Unexpected response creating config!', json=result)
+
+        c = NodeBalancerConfig(self._client, result['id'], self.id, result)
+        return c
