--- linode_api/objects/nodebalancer/__init__.py
@@ -1,3 +1,0 @@
-from .nodebalancer import NodeBalancer
-from .config import NodeBalancerConfig
-from .node import NodeBalancerNode
